requirements.txt
virtualenv==16.7.5
django==2.2.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
